1.) Changed the trial_prediction_2/model.py countries class (destination_km2 field) type to FloatField 2.) Created the load_countries script to load the countries.csv

+++ trial_prediction_2/models.py
@@ -10,7 +10,7 @@ class countries(models.Model):
     lat_destination = models.FloatField()
     lng_destination = models.FloatField()
     distance_km = models.FloatField()
-    destination_km2 = models.IntegerField()
+    destination_km2 = models.FloatField()
     destination_language = models.CharField(max_length=3)
     language_levenshtein_distance = models.FloatField()
